# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartCity OS Design Tools API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EngagementStatus =
  (typeof EngagementStatus)[keyof typeof EngagementStatus];

export const EngagementStatus = {
  active: "active",
  on_hold: "on_hold",
  archived: "archived",
} as const;

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType];

export const ProjectType = {
  new_build: "new_build",
  renovation: "renovation",
  addition: "addition",
  tenant_improvement: "tenant_improvement",
  other: "other",
} as const;

export type GeocodeSource = (typeof GeocodeSource)[keyof typeof GeocodeSource];

export const GeocodeSource = {
  nominatim: "nominatim",
  manual: "manual",
} as const;

export interface Geocode {
  latitude: number;
  longitude: number;
  jurisdictionCity: string | null;
  jurisdictionState: string | null;
  jurisdictionFips: string | null;
  source: GeocodeSource;
  geocodedAt: string;
}

export interface Site {
  address: string | null;
  geocode: Geocode | null;
  projectType: ProjectType | null;
  zoningCode: string | null;
  lotAreaSqft: number | null;
}

export interface SnapshotSummary {
  id: string;
  engagementId: string;
  engagementName: string;
  projectName: string;
  sheetCount: number | null;
  roomCount: number | null;
  levelCount: number | null;
  wallCount: number | null;
  receivedAt: string;
}

export interface EngagementSummary {
  id: string;
  name: string;
  jurisdiction: string | null;
  address: string | null;
  status: EngagementStatus;
  createdAt: string;
  updatedAt: string;
  snapshotCount: number;
  latestSnapshot: SnapshotSummary | null;
  site: Site;
}

export interface EngagementDetail {
  id: string;
  name: string;
  jurisdiction: string | null;
  address: string | null;
  status: EngagementStatus;
  createdAt: string;
  updatedAt: string;
  snapshotCount: number;
  latestSnapshot: SnapshotSummary | null;
  snapshots: SnapshotSummary[];
  site: Site;
  warnings?: string[];
}

export interface SnapshotPayload {
  projectName: string;
  [key: string]: unknown;
}

export interface SnapshotReceipt {
  id: string;
  receivedAt: string;
  engagementId: string;
  engagementName: string;
  autoCreated: boolean;
}

export type SnapshotDetailPayload = { [key: string]: unknown };

export interface SheetSummary {
  id: string;
  snapshotId: string;
  engagementId: string;
  sheetNumber: string;
  sheetName: string;
  viewCount: number | null;
  revisionNumber: string | null;
  revisionDate: string | null;
  thumbnailWidth: number;
  thumbnailHeight: number;
  fullWidth: number;
  fullHeight: number;
  sortOrder: number;
  createdAt: string;
}

export interface SnapshotDetail {
  id: string;
  engagementId: string;
  engagementName: string;
  projectName: string;
  sheetCount: number | null;
  roomCount: number | null;
  levelCount: number | null;
  wallCount: number | null;
  receivedAt: string;
  payload: SnapshotDetailPayload;
  sheets: SheetSummary[];
}

export interface SheetUploadResponse {
  uploaded: number;
  skipped: number;
  failed: number;
  errors: string[];
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  engagementId: string;
  question: string;
  history?: ChatMessage[];
  referencedSheetIds?: string[];
  /** Optional list of code atom UUIDs the user has explicitly attached
to this question. Capped server-side. Atoms are fetched, scoped
to the engagement's jurisdiction, and injected into the system
prompt alongside any retrieval-discovered atoms.
 */
  referencedAtomIds?: string[];
}

export interface ErrorResponse {
  error: string;
}

export interface ChatErrorResponse {
  error: string;
  message?: string;
}

export type JurisdictionSummaryBooksItem = {
  codeBook: string;
  edition: string;
  label: string;
  sourceName: string;
  atomCount: number;
};

export interface JurisdictionSummary {
  key: string;
  displayName: string;
  atomCount: number;
  embeddedCount: number;
  lastFetchedAt: string | null;
  books: JurisdictionSummaryBooksItem[];
}

export interface CodeAtomSummary {
  id: string;
  jurisdictionKey: string;
  codeBook: string;
  edition: string;
  sectionNumber: string | null;
  sectionTitle: string | null;
  sourceName: string;
  sourceUrl: string;
  embedded: boolean;
  fetchedAt: string;
  bodyPreview: string;
}

export type CodeAtomDetailMetadata = { [key: string]: unknown } | null;

export type CodeAtomDetail = CodeAtomSummary & {
  body: string;
  bodyHtml: string | null;
  parentSection: string | null;
  embeddingModel: string | null;
  metadata: CodeAtomDetailMetadata;
};

export type WarmupResultDrained = {
  picked: number;
  completed: number;
  failed: number;
  atomsWritten: number;
};

export interface WarmupResult {
  jurisdictionKey: string;
  enqueued: number;
  skipped: number;
  drained: WarmupResultDrained;
}

export type UpdateEngagementBody = {
  name?: string;
  address?: string;
  jurisdiction?: string;
  status?: EngagementStatus;
  projectType?: ProjectType;
  zoningCode?: string;
  lotAreaSqft?: number | null;
};

export type UploadSnapshotSheetsBody = {
  /** JSON-encoded array of SheetMetadata */
  metadata: string;
};

export type ListJurisdictionAtomsParams = {
  /**
   * @maximum 200
   */
  limit?: number;
};