# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartCity OS Design Tools API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /codes/jurisdictions/{key}/atoms:
    get:
      operationId: listJurisdictionAtoms
      tags: [codes]
      summary: List atoms for a jurisdiction (paginated, newest first)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Atom summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodeAtomSummary"
        "404":
          description: Unknown jurisdiction key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /codes/atoms/{id}:
    get:
      operationId: getCodeAtom
      tags: [codes]
      summary: Get a single code atom (full body + provenance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Atom detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeAtomDetail"
        "404":
          description: Atom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /codes/warmup/{key}:
    post:
      operationId: warmupJurisdiction
      tags: [codes]
      summary: Trigger discovery + queue drain for a jurisdiction
      description: |
        Idempotent. Inserts new TOC entries into the fetch queue (existing
        entries are skipped) and synchronously drains a small batch so the
        caller can confirm the pipeline is alive. Long-running fetches
        continue in the background queue worker.
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warmup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarmupResult"
        "404":
          description: Unknown jurisdiction key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CodeAtomSummary:
      type: object
      properties:
        id:
          type: string
        jurisdictionKey:
          type: string
        codeBook:
          type: string
        edition:
          type: string
        sectionNumber:
          type: string
          nullable: true
        sectionTitle:
          type: string
          nullable: true
        sourceName:
          type: string
        sourceUrl:
          type: string
        embedded:
          type: boolean
        fetchedAt:
          type: string
          format: date-time
        bodyPreview:
          type: string
      required:
        - id
        - jurisdictionKey
        - codeBook
        - edition
        - sectionNumber
        - sectionTitle
        - sourceName
        - sourceUrl
        - embedded
        - fetchedAt
        - bodyPreview
    CodeAtomDetail:
      allOf:
        - $ref: "#/components/schemas/CodeAtomSummary"
        - type: object
          properties:
            body:
              type: string
            bodyHtml:
              type: string
              nullable: true
            parentSection:
              type: string
              nullable: true
            embeddingModel:
              type: string
              nullable: true
            metadata:
              type: object
              additionalProperties: true
              nullable: true
          required:
            - body
            - bodyHtml
            - parentSection
            - embeddingModel
            - metadata
    WarmupResult:
      type: object
      properties:
        jurisdictionKey:
          type: string
        enqueued:
          type: integer
        skipped:
          type: integer
        drained:
          type: object
          properties:
            picked:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
            atomsWritten:
              type: integer
          required:
            - picked
            - completed
            - failed
            - atomsWritten
      required:
        - jurisdictionKey
        - enqueued
        - skipped
        - drained