# lib/api-zod/src/generated/types/codeAtomSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartCity OS Design Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeAtomSummary {
  id: string;
  jurisdictionKey: string;
  codeBook: string;
  edition: string;
  sectionNumber: string | null;
  sectionTitle: string | null;
  sourceName: string;
  sourceUrl: string;
  embedded: boolean;
  fetchedAt: Date;
  bodyPreview: string;
}