# lib/api-zod/src/generated/types/warmupResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartCity OS Design Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { WarmupResultDrained } from "./warmupResultDrained";

export interface WarmupResult {
  jurisdictionKey: string;
  enqueued: number;
  skipped: number;
  drained: WarmupResultDrained;
}